mab: add ability to locate blob in physical space.
Add a new argument -p <paddr> which takes the physical address of the
final location of the blob in physical memory.

If this is supplied mab will then fixup the addresses in the tables to
point to correct physical locations and further fixups in the kernel are
no longer needed.

This produces table blobs that are more like what we can expect from
firmware.

Signed-off-by: Graeme Gregory <graeme.gregory@linaro.org>
Acked-by: Al Stone <al.stone@linaro.org>